Revert "Use npm"
