--- index.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="description" content="Create perfect teams for foosball and other team sports with KICKET - the ultimate app.">
-    <meta name="keywords" content="team generator, foosball, table soccer">
-    <meta name="author" content="Thomas Boehringer">
-    <link rel="canonical" href="https://kicket.tboehringer.de/">
-    <meta property="og:title" content="KICKET">
-    <meta property="og:description" content="Create perfect teams for foosball and other team sports with KICKET - the ultimate app.">
-    <meta property="og:image" content="https://kicket.tboehringer.de/assets/img/kicket-logo.png">
-    <meta property="og:url" content="https://kicket.tboehringer.de/">
-    <link rel="stylesheet" href="./src/assets/css/kicket.css">
-    <title>KICKET</title>
-    <link rel="apple-touch-icon" sizes="76x76" href="src/assets/img/favicons/apple-touch-icon.png">
-    <link rel="icon" type="image/png" sizes="32x32" href="src/assets/img/favicons/favicon-32x32.png">
-    <link rel="icon" type="image/png" sizes="16x16" href="src/assets/img/favicons/favicon-16x16.png">
-    <link rel="manifest" href="src/assets/img/favicons/site.webmanifest">
-    <link rel="mask-icon" href="src/assets/img/favicons/safari-pinned-tab.svg" color="#366164">
-    <meta name="msapplication-TileColor" content="#366164">
-    <meta name="theme-color" content="#366164">
-  </head>
-  <body>
-    <noscript>
-      <meta http-equiv="refresh" content="0.0;url=nojs/index.html">
-    </noscript>
-    <div id="app"></div>
-    <script type="module" src="/src/main.js"></script>
-  </body>
-</html>
