+++ src/assets/css/kicket.css
@@ -3,8 +3,11 @@ html {
     min-height: 100%;
 }
 body {
-  background-color: #366164 !important;
+  background-color: #366164;
   margin-bottom: 45px;
+}
+.navbar-dark {
+  background-color: #366164;
 }
 .btn-success, .btn-primary {
   background-color: #16a085;
@@ -17,4 +20,12 @@ body {
 .btn-success:disabled, .btn-primary:disabled {
   background-color: #48c9b0;
   border-color: #48c9b0 ;
+}
+.footer {
+  position: absolute;
+  bottom: 0;
+  width: 100%;
+  height: 40px;
+  line-height: 40px;
+  background-color: #53BBAD;
 }
