+++ src/components/Navbar.js
@@ -1,0 +1,8 @@
+app.component('knavbar', {
+    template:
+    /*html*/
+    `
+    <nav class="navbar navbar-dark py-0">
+        <a class="navbar-brand py-0" href="index.html"><img src="./assets/img/kicket-logo-nav.png" height="30" alt="KICKET"></a>    
+    </nav>`
+})
