--- src/components/Navbar.vue
@@ -1,19 +1,0 @@
-<script setup>
-</script>
-
-<template>
-  <nav class="navbar navbar-dark py-0">
-    <div class="container-fluid">
-      <a class="navbar-brand py-0" href="index.html"><img src="../assets/img/kicket-logo-nav.png" height="30" alt="KICKET"></a>    
-   </div>
-  </nav>
-</template>
-
-<style scoped>
-nav {
-  padding: .5rem 1rem
-}
-.navbar-dark {
-  background-color: #366164;
-}
-</style>
