+++ src/components/TeamList.js
@@ -1,6 +1,6 @@
-<script setup>
-defineProps({
-  started: {
+app.component('team-list', {
+  props: {
+    started: {
       type: Boolean,
       required: true
     },
@@ -9,31 +9,16 @@ defineProps({
       required: true
     },
     players: {
-      type: Array,
-    }
-})
-</script>
-<script>
-export default {
-  methods: {
-    restart() {
-      this.$emit('restart-game')
+      type: Array
     }
   },
-  computed: {
-    getSumOfPlayers() {
-      return this.players.length
-    }
-  }
-}
-</script>
-
-<template>
-  <div id="teams" v-if="started">
+  template:
+    /*html*/
+    `<div id="teams" v-if="started">
     <h2>Teams for {{ getSumOfPlayers }} players</h2>
     <table id="teamList" class="table table-bordered table-hover">
       <thead>
-        <tr class="table-dark">
+        <tr class="thead-dark">
           <th scope="col">Team</th>
           <th scope="col">Defense</th>
           <th scope="col">Offense</th>
@@ -47,9 +32,15 @@ export default {
       </tbody>
     </table>
     <button class="btn btn-dark" v-on:click="restart">Back</button>
-  </div>
-</template>
-
-<style>
-
-</style>
+  </div>`,
+  methods: {
+    restart() {
+      this.$emit('restart-game')
+    }
+  },
+  computed: {
+    getSumOfPlayers() {
+      return this.players.length;
+    }
+  }
+})
