+++ src/index.html
@@ -1,0 +1,69 @@
+<!Doctype html>
+<html lang="de">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <meta name="description" content="Create perfect teams for foosball and other team sports with KICKET - the ultimate app.">
+    <meta name="keywords" content="team generator, foosball, table soccer">
+    <meta name="author" content="Thomas Boehringer">
+    <link rel="canonical" href="https://kicket.tboehringer.de/">
+    <meta property="og:title" content="KICKET">
+    <meta property="og:description" content="Create perfect teams for foosball and other team sports with KICKET - the ultimate app.">
+    <meta property="og:image" content="https://kicket.tboehringer.de/assets/img/kicket-logo.png">
+    <meta property="og:url" content="https://kicket.tboehringer.de/">
+    <title>KICKET</title>
+    <script src="https://unpkg.com/vue@3/dist/vue.global.js"></script>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+    <link rel="stylesheet" href="./assets/css/kicket.css">
+    <!-- favicon/icons-->
+    <link rel="apple-touch-icon" sizes="76x76" href="assets/img/favicons/apple-touch-icon.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="assets/img/favicons/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="assets/img/favicons/favicon-16x16.png">
+    <link rel="manifest" href="assets/img/favicons/site.webmanifest">
+    <link rel="mask-icon" href="assets/img/favicons/safari-pinned-tab.svg" color="#366164">
+    <meta name="msapplication-TileColor" content="#366164">
+    <meta name="theme-color" content="#366164">
+  </head>
+  <body>
+    <noscript>
+      <meta http-equiv="refresh" content="0.0;url=nojs/index.html">
+    </noscript>
+    <div id="app">
+      <knavbar></knavbar>
+      <main role="main" class="container">
+        <h1 class="mt-3 text-center">{{ name }}</h1>
+        <p class="lead text-center">{{ description }}</p>
+
+        <game-configuration
+           :started="started"
+           @start-game="startGame">
+        </game-configuration>
+        <team-list
+          :started="started"
+          :teams="generateTeams()"
+          :players="getPlayers"
+          @restart-game="restartGame">
+        </team-list>
+      </main>
+
+      <kfooter
+        :copyright="getCopyright"
+        :gdpr="getGDPR">
+      </kfooter>
+    </div>
+
+    <script src="./main.js"></script>
+    <script src="./components/Navbar.js"></script>
+    <script src="./components/Footer.js"></script>
+    <script src="./components/GameConfiguration.js"></script>
+    <script src="./components/TeamList.js"></script>
+
+    <script>
+      const kicket = app.mount('#app')
+    </script>
+
+    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
+  </body>
+</html>
