+++ src/nojs/index.html
@@ -13,12 +13,12 @@
     <meta property="og:url" content="https://kicket.tboehringer.de/">
     <title>KICKET</title>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-    <link rel="stylesheet" href="../../src/assets/css/kicket.css">
-    <link rel="apple-touch-icon" sizes="76x76" href="../../src/assets/img/favicons/apple-touch-icon.png">
-    <link rel="icon" type="image/png" sizes="32x32" href="../../src/assets/img/favicons/favicon-32x32.png">
-    <link rel="icon" type="image/png" sizes="16x16" href="../../src/assets/img/favicons/favicon-16x16.png">
-    <link rel="manifest" href="../../src/assets/img/favicons/site.webmanifest">
-    <link rel="mask-icon" href="../../src/assets/img/favicons/safari-pinned-tab.svg" color="#366164">
+    <link rel="stylesheet" href="../assets/css/kicket.css">
+    <link rel="apple-touch-icon" sizes="76x76" href="../assets/img/favicons/apple-touch-icon.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="../assets/img/favicons/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="../assets/img/favicons/favicon-16x16.png">
+    <link rel="manifest" href="../assets/img/favicons/site.webmanifest">
+    <link rel="mask-icon" href="../assets/img/favicons/safari-pinned-tab.svg" color="#366164">
     <meta name="msapplication-TileColor" content="#366164">
     <meta name="theme-color" content="#366164">
   </head>
